LayoutTests/imported/w3c/web-platform-tests/cssom-view/interfaces-expected.txt

CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported

CSSOM View IDL tests


FAIL Test driver promise_test: Unhandled rejection with value: object "Unrecognised tokens, line 1 (tokens: '{"error": {')
[
    {
        "type": "other",
        "value": "{"
    },
    {
        "type": "string",
        "value": "\"error\""
    },
    {
        "type": "other",
        "value": ":"
    },
    {
        "type": "whitespace",
        "value": " "
    },
    {
        "type": "other",
        "value": "{"
    }
]"
